Switch C code to use int not long to be consistent with the Java code.

+++ pi_c_mpi.c
@@ -1,7 +1,7 @@
 /*
  *  A C program to calculate Pi using quadrature.  This is an SPMD realization using OpenMPI.
  *
- *  Copyright © 2008-2011 Russel Winder
+ *  Copyright © 2008--2011 Russel Winder
  */
 
 #include <stdio.h>
@@ -9,18 +9,18 @@
 #include "microsecondTime.h"
 
 int main ( int ac , char ** av ) { // MPI requires write access to these parameters!
-  const long n = 1000000000l ;
+  const int n = 1000000000 ;
   const double delta = 1.0 / n ;
   int nProcessors , myId ;
   const long long startTimeMicros = microsecondTime ( ) ;
   MPI_Init ( &ac , &av ) ;
   MPI_Comm_size ( MPI_COMM_WORLD , &nProcessors ) ;
   MPI_Comm_rank ( MPI_COMM_WORLD , &myId ) ;
-  const long sliceSize = n / nProcessors ;
-  const long start = 1 + myId * sliceSize ;
-  const long end = ( myId + 1 ) * sliceSize ;
+  const int sliceSize = n / nProcessors ;
+  const int start = 1 + myId * sliceSize ;
+  const int end = ( myId + 1 ) * sliceSize ;
   double localSum = 0.0 ;
-  for ( long i = start ; i <= end ; ++i ) {
+  for ( int i = start ; i <= end ; ++i ) {
     const double x = ( i - 0.5 ) * delta ;
     localSum += 1.0 / ( 1.0 + x * x ) ;
   }
@@ -31,7 +31,7 @@ int main ( int ac , char ** av ) { // MPI requires write access to these paramet
     const double pi = 4.0 * sum * delta ;
     const double elapseTime = ( microsecondTime ( ) - startTimeMicros ) / 1e6 ;
     printf ( "==== C MPI pi = %.18lf\n" , pi ) ;
-    printf ( "==== C MPI iteration count = %ld\n" , n ) ;
+    printf ( "==== C MPI iteration count = %d\n" , n ) ;
     printf ( "==== C MPI elapse time = %lf\n" , elapseTime ) ;
     printf ( "==== C MPI processor count = %d\n" ,  nProcessors ) ;
   }
